wip: change structure
